functions/requirements.txt
# Appwrite Functions Requirements
appwrite==12.0.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9

# AI/ML (Lightweight for Appwrite)
requests==2.31.0
tenacity==8.2.3

# Document Processing (Basic)
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
